check for invalid signatures
